Change storage of MATLAB variable dimensions to a size_t
* src/matio.h (matvar_t): Change type of dims field to size_t pointer

* src/mat.c,src/matio.h (Mat_VarCreate): Change type of 'dims' argument
      to size_t pointer

* src/mat5.c,src/mat5.h (WriteDataSlab2,WriteCharDataSlab2): Change type
      of 'dims' argument to size_t pointer

* src/read_data.c,src/matio_private.h (ReadDataSlab2,ReadDataSlabN,
  ReadCompressedDataSlab2,ReadCompressedDataSlabN): Changed type of
      'dims' argument to size_t pointer

* test/test_mat.c (test_write,test_write_compressed,
  test_write_complex_compressed,test_readvar4,
  test_write_compressed_struct,test_write_cell,
  test_write_compressed_cell,test_write_null,test_writeslab,
  test_writenan,test_writeinf,test_writesparse,
  test_write_compressed_sparse): Changed type of 'dims' variable to
      size_t

+++ src/mat.c
@@ -441,7 +441,7 @@ Mat_VarCalloc(void)
  */
 matvar_t *
 Mat_VarCreate(const char *name,enum matio_classes class_type,
-    enum matio_types data_type,int rank,int *dims,void *data,int opt)
+    enum matio_types data_type,int rank,size_t *dims,void *data,int opt)
 {
     int i, nmemb = 1, nfields = 0;
     matvar_t *matvar = NULL;
